backend/requirements.txt
Flask
python-dotenv
requests
pdfplumber
python-docx
flask-cors
